+++ CHANGELOG.rst
@@ -20,6 +20,8 @@ see TODO.md for list of TODO items
 
   * modified get_menu example for new menu structure
 
+  * added elem_validators to get_list to validate list items
+
 * v0.2.12:
 
   * renamed ChoicesValidator to ChoiceValidator